add BaseModel and move build and set_initial_state there
