+++ pybop/models/lithium_ion/__init__.py
@@ -1,6 +1,7 @@
 #
 # Import lithium ion based models
 #
+from .base_model import BaseModel
 from .sp_diffusion import SPDiffusion
 from .grouped_spm import GroupedSPM
 from .grouped_spme import GroupedSPMe
